manytask/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "manytask"
dynamic = ["version"]
description = "Manytask - a web app to organize courses and collect students scores"
readme = "README.md"
requires-python = ">=3.12,<4.0"
license = {text = "MIT"}
authors = [
    {name = "Manytask Team"}
]
dependencies = [
    "alembic==1.18.3",
    "authlib==1.6.12",
    "gunicorn==25.1.0",
    "flask==3.1.3",
    "flask-wtf>=1.3.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.13.3",
    "python-dotenv==1.2.2",
    "python-gitlab==8.3.0",
    "pyyaml==6.0.1",
    "requests==2.33.0",
    "sqlalchemy>=2.0.44,<3.0.0",
    "werkzeug==3.1.6",
    "httpx>=0.28.1",
    "yandexcloud>=0.388.0",
    "grpcio>=1.78.0",
]


[project.optional-dependencies]
dev = [
    "ruff==0.15.4",
    "mypy==1.20.0",
    "types-requests==2.33.0.20260327",
    "types-PyYAML==6.0.12.12",
    "pytest==9.0.3",
    "pytest-cov==7.1.0",
    "pytest-mock>=3.0.0,<4.0.0",
    "requests-mock>=1.0.0,<2.0.0",
    "testcontainers==4.14.0",
    "GitPython>=3.0.0,<4.0.0",
    "bs4>=0.0.2",
    "pytest-xdist>=3.8.0",
]

[tool.mypy]
no_incremental = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_any_generics = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
allow_untyped_decorators = true
exclude = [
    "tests/",
    ".tmp/",
    ".venv/",
    "manytask/migrations/versions",
    "checker/tests/",
    "checker/.tmp/"
]


[tool.pyright]
include = ["manytask", "checker/checker"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    ".venv",
    "venv",
    ".tmp",
    "tests",
    "checker/tests",
    "checker/.tmp"
]


[tool.pytest.ini_options]
minversion = "6.0"
python_files = "test_*.py"
addopts = "--cov=manytask/ --cov-report=term-missing tests/"
testpaths = [
    "tests",
    ".tmp"
]


[tool.ruff]
line-length = 120
target-version = "py312"
src = ["manytask", "tests", "checker/checker"]

# Exclude patterns
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    "*.ipynb",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    ".tmp",
    "manytask/migrations/versions",
    "checker/tests"
]

[tool.ruff.lint]
# Combine both configurations' rules
select = ["E", "F", "I", "TID252", "TID253", "C901", "PLR"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.mccabe]
# Maximum cyclomatic complexity allowed
max-complexity = 15

[tool.ruff.lint.pylint]
max-nested-blocks = 5
max-args = 7

[tool.ruff.lint.isort]
known-first-party = ["manytask", "tests", "checker"]

[tool.setuptools]
packages = ["manytask"]

[tool.setuptools.dynamic]
version = {file = "VERSION"}
